get-weather-api-function/requirements.txt
google-cloud-secret-manager==2.20.2
google-cloud-core==2.4.1
google-cloud-storage==2.18.2